+++ dashboard/views/automated-embed/player.php
@@ -19,7 +19,7 @@ $info_card = [
 
 ?>
 
-<form action="<?php echo get_admin_url() . 'admin.php?page=dm-automated-embed-settings&tab=player&action=save_data'; ?>" method="post">
+<form action="<?php echo get_admin_url() . 'admin.php?page=dm-automated-embed-settings&tab=player&action=save_data'; ?>" method="post" id="automated-embed-form">
 
     <?php wp_nonce_field("dm_save_data", "dm_save_data", true);?>